refactored hoursOfOperation component to clean up component file and refactored code
